67314155 test

+++ src/main/scala/job/remote/flashbuy/Cid2Item.scala
@@ -34,21 +34,15 @@ object Cid2Item extends RemoteSparkJob {
                |""".stripMargin).rdd.map(row => {
             val cate3Id_geohash = row.getAs[String](0)
             val sku_id = row.getAs[Long](1)
-            val cnt = row.getAs[Long](2)
+            val cnt = row.getAs[Double](2)
             (cate3Id_geohash, (sku_id, cnt))
         }).groupByKey.mapValues(iter => {
             val entities = iter.toArray.sortBy(_._2).takeRight(100)
-            val factors = ArrayOperations.softmax(entities.map(_._2.toDouble))
+            val factors = ArrayOperations.softmax(entities.map(_._2))
             val results = entities.map(_._1).zip(factors).map(x=>s"${x._1}:${x._2}")
             results
         }).toDF("key", "value")
         val partition = Map("date" -> dt, "branch" -> "cid", "method" -> "pt_cid_sales_sku_base")
         saveAsTable(spark, df, "recsys_linshou_multi_recall_results_v2", partition=partition)
-
-
-
-
     }
-
-
 }
